+++ nextgen/config/src/inherited_tasks_config.rs
@@ -1,6 +1,6 @@
 use crate::language_platform::{LanguageType, PlatformType};
 use crate::project::{validate_deps, TaskConfig};
-use crate::project_config::{validate_tasks, ProjectType};
+use crate::project_config::ProjectType;
 use crate::shapes::InputPath;
 use crate::validate::check_yml_extension;
 use moon_common::path::standardize_separators;
@@ -51,7 +51,7 @@ cacheable!(
         #[setting(merge = merge::append_vec)]
         pub implicit_inputs: Vec<InputPath>,
 
-        #[setting(nested, merge = merge::merge_btreemap, validate = validate_tasks)]
+        #[setting(nested, merge = merge::merge_btreemap)]
         pub tasks: BTreeMap<Id, TaskConfig>,
     }
 );