+++ nextgen/task-builder/src/tasks_builder_error.rs
@@ -1,0 +1,14 @@
+use miette::Diagnostic;
+use moon_common::{Id, Style, Stylize};
+use thiserror::Error;
+
+#[derive(Error, Debug, Diagnostic)]
+pub enum TasksBuilderError {
+    #[diagnostic(code(task_builder::unknown_extends))]
+    #[error(
+        "Task {} is extending an unknown task {}.",
+        .source_id.style(Style::Id),
+        .target_id.style(Style::Id),
+    )]
+    UnknownExtendsSource { source_id: Id, target_id: Id },
+}